backend/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings
httpx
beautifulsoup4
python-dotenv
firebase-admin
